Structure /create-server to allow for importing in future
