+++ apps/web/app/(with-sidebar)/create-server/new/page.module.css
@@ -1,0 +1,9 @@
+.layout {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+	margin-right: auto;
+	margin-left: auto;
+	width: 100%;
+	max-width: 800px;
+}
