+++ apps/web/app/(with-sidebar)/create-server/page.module.css
@@ -2,8 +2,25 @@
 	display: flex;
 	flex-direction: column;
 	gap: 1rem;
-	margin-right: auto;
-	margin-left: auto;
 	width: 100%;
 	max-width: 800px;
+}
+
+.createOption {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	gap: 8px;
+	width: 200px;
+	height: 200px;
+	font-size: var(--font-size-large);
+}
+
+.createOptions {
+	display: flex;
+	flex-direction: row;
+	gap: 1rem;
+	padding: 8px;
+	width: 100%;
 }
